+++ NEWS.md
@@ -1,6 +1,15 @@
 # <Development version: cSEM 0.4.0.9000
 
-- 
+- Removed argument `.verbose` from `testOMF()` as it did not have any effect (#445).
+
+- Bug fix: GSCAm (i.e., `.approach_weights = "GSCA"` with constructs modeled as 
+  common factors) no longer fails when a single indicator construct is supplied (#441)
+  
+- Bug fix: `predict()` no longer fails when LOOCV is used (#337)
+
+- Bug fix: fix print method for `summarize()` when resampling with constant values
+  (weights or loadings) is conducted. The standard error, t-value, p-value and
+  CI are properly set to `NA` now. (#433)
 
 # cSEM 0.4.0 (2021-04-20)